pyproject.toml
[tool.poetry]
name = "gov-context-llm"
version = "0.1.0"
description = ""
authors = ["Michael Kalish <michael.kalish@focusconsulting.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
tensorflow = "^2.16.2"
transformers = "^4.42.3"
tf-keras = "^2.16.0"
langchain = "^0.2.6"
vtt-to-srt3 = "^0.2.0.1"
langchain-community = "^0.2.6"
pysrt = "^1.1.2"
openai = "^1.35.10"
chromadb = "^0.5.3"
langchain-openai = "^0.1.14"
langchainhub = "^0.1.20"
lark = "^1.1.9"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
streamlit = "^1.36.0"
watchdog = "^4.0.1"
pypdf = "^4.2.0"


[tool.poetry.group.dev.dependencies]
types-beautifulsoup4 = "^4.12.0.20240511"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
